+++ engine/src/main/java/dev/gamekit/animation/Animation.java
@@ -1,7 +1,6 @@
 package dev.gamekit.animation;
 
 import dev.gamekit.core.Application;
-import dev.gamekit.core.Constants;
 import dev.gamekit.utils.ValueCallback;
 
 import static dev.gamekit.utils.Math.clamp;
@@ -113,7 +112,7 @@ public class Animation {
   /** Called internally by the application game loop to update this animation */
   public void update() {
     if (state == State.RUNNING) {
-      value = clamp(value + 0.001 * rate * Constants.FRAME_INTERVAL_MS, 0, 1);
+      value = clamp(value + 0.001 * rate * Application.FRAME_INTERVAL_MS, 0, 1);
 
       if (valueListener != null)
         valueListener.run(getValue());