+++ CHANGELOG.md
@@ -52,7 +52,8 @@
 - Added `ValueCallback<T>` class which is a single abstract method interfaces whose `run` method
   accepts a single variable of type `T`
 - Added state machine package to engine module
-- Added `static Vector sum(Vector...)` which returns a vector summing all input vectors
+- Added `static Vector sum(Vector...)` to `Vector` which returns a vector summing all input vectors
+- Added `static double lerpAngle(double, double, double)` to `Math` class
 
 ### Changed